requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.21.0
airportsdata==20250909
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.8.0
apache-tvm-ffi==0.1.0b15
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.1.0
babel==2.17.0
backoff==2.2.1
banks==2.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
blis==1.3.0
blobfile==3.0.0
Brotli==1.0.9
build==1.3.0
cachetools==5.5.2
captum==0.8.0
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.3.2
class-registry==2.1.2
click==8.1.8
cloudpathlib==0.21.1
cloudpickle==3.1.1
cmudict==1.0.32
colorama==0.4.6
comm==0.2.2
compressed-tensors==0.12.2
confection==0.1.5
contourpy==1.3.1
cuda-bindings==13.0.2
cuda-pathfinder==1.3.1
cuda-python==13.0.2
curl_cffi==0.11.4
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
datasets==3.3.2
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
einops==0.8.1
en_core_web_sm==3.8.0
et_xmlfile==2.0.0
evaluate==0.4.3
executing==2.2.0
fastapi==0.119.1
fastjsonschema==2.21.1
filelock==3.13.1
filetype==1.2.0
flashinfer-python==0.4.0
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.10.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.0
greenlet==3.2.4
griffe==1.14.0
grpcio==1.75.1
grpcio-reflection==1.75.1
grpcio-status==1.70.0
grpcio-tools==1.75.1
h11==0.14.0
h5py==3.13.0
hf_transfer==0.1.9
hf-xet==1.1.9
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.7
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
interegular==0.3.3
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
joblib==1.5.0
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter_client==8.6.3
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
langchain==0.3.27
langchain-core==0.3.76
langchain-text-splitters==0.3.11
langcodes==3.5.0
langsmith==0.4.27
language_data==1.3.0
lark==1.3.0
llama-cloud==0.1.41
llama-cloud-services==0.6.64
llama_cpp_python==0.3.16
llama-index-core==0.14.1
llama-index-embeddings-huggingface==0.6.1
llama-index-instrumentation==0.4.0
llama-index-llms-llama-cpp==0.5.1
llama-index-llms-openai==0.5.6
llama-index-workflows==2.1.0
llguidance==0.7.30
llvmlite==0.44.0
loguru==0.7.3
lxml==6.0.2
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
modelscope==1.31.0
mpmath==1.3.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.13.0
nltk==3.9.1
notebook==7.4.2
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-frontend==1.15.0
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-cutlass-dsl==4.2.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==1.99.1
openai-harmony==0.0.4
openpyxl==3.1.5
orjson==3.11.3
outlines==0.1.11
outlines_core==0.1.26
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partial-json-parser==0.2.1.1.post6
patsy==1.0.1
pexpect==4.9.0
pillow==11.0.0
pip==24.2
platformdirs==4.3.8
pluggy==1.6.0
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.0
proto-plus==1.26.0
protobuf==6.33.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-spy==0.4.1
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybase64==1.4.2
pycocotools==2.0.8
pycountry==24.6.1
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.1
pyphen==0.17.2
pyproject_hooks==1.2.0
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rouge_score==0.1.2
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
Send2Trash==1.8.3
sentence-transformers==5.1.0
sentencepiece==0.2.1
setproctitle==1.3.7
setuptools==80.9.0
sgl-kernel==0.3.15
sglang==0.5.3.post3
shap==0.47.2
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
smart-open==7.1.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
spacy==3.8.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
srsly==2.5.1
stack-data==0.6.3
starlette==0.48.0
statsmodels==0.14.5
style==1.1.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
terminado==0.18.1
textstat==0.7.7
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==1.0.16
tinycss2==1.4.0
tokenizers==0.19.1
torch==2.8.0
torch_memory_saver==0.0.9rc2
torchao==0.9.0
torchaudio==2.8.0
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.44.0
triton==3.4.0
typer==0.15.4
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.1
update==0.0.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
uv==0.9.4
uvicorn==0.38.0
uvloop==0.22.1
vqaloader==0.1.4
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.44.0
widgetsnbextension==4.0.14
wikipedia==1.4.0
word2number==1.1
wrapt==1.17.2
xgboost==3.0.1
xgrammar==0.1.25
XlsxWriter==3.2.3
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.24.0
